# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Common User App",
      "description": "This is a sample server for a book store.",
      "contact": {
        "name": "API Support",
        "url": "http://www.example.com/support",
        "email": "support@example.com"
      },
      "version": "1.0.1"
    },
    "servers": [
      {
        "url": "http://localhost:3000/api-documentation",
        "description": "test"
      }
    ],
    "paths": {
      "/user": {
        "post": {
          "description": "Essa rota é responsavel por cadastrar um usuário.",
          "tags": ["User"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserReq" },
                "examples": {
                  "Patrick": {
                    "value": {
                      "name": "Patrick Nekel",
                      "email": "nekeldev@gmail.com",
                      "password": "1234",
                      "is_adm": false
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "409": {
              "description": "Email already exists"
            },
            "201": {
              "description": "Created",
               "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/UserRes" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "Patrick Nekel",
                        "email": "nekeldev@gmail.com",
                        "is_adm": false,
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-25T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "get": {
          "description": "Essa rota é responsavel por listar todos os usuários. Apenas usuários admin podem acessar essa rota.",
          "tags": ["User"],
          "requestBody": {
            "content": {
              "application/json": {
              }
            }
          },
          "responses": {
            "401": {
              "description": "Unauthorized"
            },
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/UserList" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "Patrick Nekel",
                        "email": "nekeldev@gmail.com",
                        "is_adm": false,
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-25T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/user/{id}": {
        "get": {
          "description": "Essa rota é responsavel por listar as informações do usuário logado. Precisa de autenticação.",
          "tags": ["User"],
          "requestBody": {
            "content": {
              "application/json": {                
              }
            }   
          },
          "responses": {
            "401": {
              "description": "Missing authorization headers, Invalid Token"
            },
            "404": {
              "description": "User not found"
            },
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/UserRes" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "Patrick Nekel",
                        "email": "nekeldev@gmail.com",
                        "is_adm": false,
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-25T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "patch": {
          "description": "Essa rota é responsavel por atualizar um usuário específico. Precisa de autenticação. Um usuário não admin pode buscar apenas seu próprio usuário.",
          "tags": ["User"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserReq" },
                "examples": {
                  "Example": {
                    "value": {
                      "name": "Patrick Nekel Jr",
                      "email": "nekeldev@yahoo.com",
                      "password": "123456"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "401": {
              "description": "Missing authorization headers, Invalid Token, Missing admin permissions"
            },
            "404": {
              "description": "User not found"
            },
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/UserRes" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "Patrick Nekel Jr",
                        "email": "nekeldev@yahoo.com",
                        "is_adm": false,
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-26T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "description": "Essa rota é responsavel por excluir um usuário específico. Precisa de autenticação. Um usuário não admin pode buscar apenas seu próprio usuário.",
          "tags": ["User"],
          "requestBody": {
            "content": {
              "application/json": {               
              }
            }
          },
          "responses": {
            "401": {
              "description": "Missing authorization headers, Invalid Token, Missing admin permissions"
            },
            "404": {
              "description": "User not found"
            },
            "204": {
              "description": "OK",
              "content": {
                "application/json": {                 
                }
              }
            }
          }
        }
      },            
      "/book": {
        "post": {
          "description": "Essa rota é responsavel por cadastrar um livro. Precisa de autenticação. Apenas usuários admin têm acesso a esse endpoint.",
          "tags": ["Book"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookReq" },
                "examples": {
                  "Example": {
                    "value": {
                      "name": "A História da Programação",
                      "price": 30.49,
                      "description": "Excelente livro.",
                      "author": "Peter Jr"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "401": {
              "description": "Missing authorization headers, Invalid Token, Unauthorized. Only admin has access to this endpoint."
            },
            "201": {
              "description": "Created",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/BookRes" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "A História da Programação",
                        "price": 30.49,
                        "description": "Excelente livro.",
                        "author": "Peter Jr",
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-26T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "get": {
          "description": "Essa rota é responsavel por listar todos os livros. É possível filtrar por título e/ou autor. Qualquer pessoa tem acesso a esse endpoint (não precisa estar autenticado).",
          "tags": ["Book"],
          "requestBody": {
            "content": {
              "application/json": {
              },
              "examples": {
                "product": {
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "schema": {
                  "type": "object[]",
                  "$ref": "#/components/schemas/BookList"
                }
              }
            }
          }
        }
      },
      "/book/{id}": {
        "get": {
          "description": "Essa rota é responsavel por listar umlivro específico. Qualquer pessoa tem acesso a esse endpoint (não precisa estar autenticado).",
          "tags": ["Book"],
          "requestBody": {
            "content": {
              "application/json": {                
              }
            }   
          },
          "responses": {
            "404": {
              "description": "Book not found"
            },
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/BookRes" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "A História da Programação",
                        "price": 30.49,
                        "description": "Excelente livro.",
                        "author": "Peter Jr",
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-26T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "patch": {
          "description": "Essa rota é responsavel por atualizar um livro específico. Precisa de autenticação. Apenas um usuário admin tem acesso a essa rota.",
          "tags": ["Book"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookReq" },
                "examples": {
                  "Patrick": {
                    "value": {
                      "price": 40.49
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "401": {
              "description": "Missing authorization headers, Invalid Token, Unauthorized. Only admin has access to this endpoint."
            },
            "404": {
              "description": "Book not found"
            },
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/BookRes" },
                  "examples": {
                    "Example": {
                      "value": {
                        "name": "A História da Programação",
                        "price": 40.49,
                        "description": "Excelente livro.",
                        "author": "Peter Jr",
                        "created_at": "2022-01-25T02:13:05.384Z",
	                      "updated_at": "2022-01-26T02:13:05.384Z"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "description": "Essa rota é responsavel por excluir um livro específico. Precisa de autenticação. Apenas um usuário admin tem acesso a essa rota.",
          "tags": ["Book"],
          "requestBody": {
            "content": {
              "application/json": {               
              }
            }
          },
          "responses": {
            "401": {
              "description": "Missing authorization headers, Invalid Token, Unauthorized. Only admin has access to this endpoint."
            },
            "404": {
              "description": "User not found"
            },
            "204": {
              "description": "OK",
              "content": {
                "application/json": {                 
                }
              }
            }
          }
        }
      }, 
...
    },
    "components": {
      "schemas": {
        "UserReq": {
          "schemas":{
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "is_adm": {
                "type": "boolean"
              }
            }
          }
        },
        "UserRes": {
          "schemas": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "is_adm": {
                "type": "boolean"
              },
              "created_at": {
                "type": "timestamp"
              },
              "updated_at": {
                "type": "timestamp"
              }
            }
          }
        },
        "UserList": {
          "schemas": {
            "ArrayOfUsers": {
              "type":"array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "is_adm": {
                    "type": "boolean"
                  },
                  "created_at": {
                    "type": "timestamp"
                  },
                  "updated_at": {
                    "type": "timestamp"
                  }
                }
              }
           
            }
          }
        },
...
        "BookReq": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "price": {
              "type": "number"
            },
            "description": {
              "type": "string"
            },
            "author": {
              "type": "string"
            }
          }
        },
        "BookRes": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "price": {
              "type": "number"
            },
            "description": {
              "type": "string"
            },
            "author": {
              "type": "string"
            },
            "created_at": {
              "type": "timestamp"
            },
            "updated_at": {
              "type": "timestamp"
            }
          }
        },
...
      }
    }
  } 